Move queryCache into figbird, less layers
